+++ RELEASE-NOTES.md
@@ -23,6 +23,7 @@
 * Cleaning up the output directory after coping files to the stage for Streamlit, Notebook, SPCS Service and project.
 * Added interactive mode for `snow sql`.
 * Added support for async SQL queries (`;>` syntax).
+* Added support for `!queries`, `!result` and `!abort` commands from SnowSQL.
 
 ## Fixes and improvements